// packages/connect/ic/providers/types.ts
type CustomError<T> = { kind: T; message?: string };

export enum CreateActorError {
    FetchRootKeyFailed = 'FETCH_ROOT_KEY_FAILED',
    CreateActorFailed = 'CREATE_ACTOR_FAILED',
    NotInitialized = 'NOT_INITIALIZED',
    LocalActorsNotSupported = 'LOCAL_ACTORS_NOT_SUPPORTED',
}

export type CreateActorResult<Service> = Result<
    ActorSubclass<Service>,
    CustomError<CreateActorError>
>;

export enum ConnectError {
    NotInitialized = 'NOT_INITIALIZED',
    NotInstalled = 'NOT_INSTALLED',
    ConnectFailed = 'CONNECT_FAILED',
    IsLocked = 'IS_LOCKED',
}

export type ConnectResult = Result<boolean, CustomError<ConnectError>>;

export enum DisconnectError {
    DisconnectFailed = 'DISCONNECT_FAILED',
    NotInitialized = 'NOT_INITIALIZED',
}

export type DisconnectResult = Result<boolean, CustomError<DisconnectError>>;

export enum InitError {
    NotInstalled = 'NOT_INSTALLED',
    InitFailed = 'INIT_FAILED',
    FetchRootKeyFailed = 'FETCH_ROOT_KEY_FAILED',
}

export type InitResult = Result<{ isConnected: boolean }, CustomError<InitError>>;

export interface IConnector {
    init: () => Promise<InitResult>;
    config: any;
    meta: {
        features: Array<string>;
        icon: {
            light: string;
            dark: string;
        };
        id: string;
        name: string;
    };
    isConnected: () => Promise<boolean>;
    createActor: <Service>(
        canisterId: string,
        interfaceFactory: IDL.InterfaceFactory,
        // config?: {},
    ) => Promise<CreateActorResult<Service>>;
    connect: (options?: { delegationModes: Array<string> }) => Promise<ConnectResult>;
    disconnect: () => Promise<DisconnectResult>;
    principal?: string;
}